Factor out CollectionTesting utilities.
