+++ Tests/TensorFlowTests/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_library(TensorFlowTests
+  CollectionTesting.swift
   ContextTests.swift
   EpochsTests.swift
   FreezableTests.swift
@@ -17,6 +18,7 @@ add_library(TensorFlowTests
   LossTests.swift
   OptimizerTests.swift
   RuntimeTests.swift
+  SamplingTests.swift
   SequencedTests.swift
   SequencedTests.swift.gyb
   SequentialTests.swift
